Update time filter (tomorrow -> nextBusinessDay) update dateUtil method to static LeaveRequest CRUD session check

+++ src/main/java/com/advantech/servlet/admin/DeleteLeaveReq.java
@@ -18,21 +18,32 @@ import javax.servlet.http.*;
  */
 @WebServlet(name = "DeleteLeaveReq", urlPatterns = {"/DeleteLeaveReq"})
 public class DeleteLeaveReq extends HttpServlet {
-
+    
+    private int SYTEM_MANAGER_PERMISSION;
+    
+    @Override
+    public void init() throws ServletException {
+        SYTEM_MANAGER_PERMISSION = StringParser.strToInt(getServletContext().getInitParameter("SYTEM_MANAGER_PERMISSION"));
+    }
+    
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
         res.sendError(HttpServletResponse.SC_FORBIDDEN);
     }
-
+    
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse res)
             throws ServletException, IOException {
-
+        
         PrintWriter out = res.getWriter();
         String id = req.getParameter("id");
-        out.print(deleteLeaveRequest(StringParser.strToInt(id)));
+        if ((int) req.getSession(false).getAttribute("permission") >= SYTEM_MANAGER_PERMISSION) {
+            out.print(deleteLeaveRequest(StringParser.strToInt(id)));
+        } else {
+            out.print(false);
+        }
     }
-
+    
     private boolean deleteLeaveRequest(int id) {
         return BasicService.getLeaveRequestService().deleteLeaveRequest(id);
     }
